requirements.txt
discord.py
apscheduler
pytz
parsedatetime
unyt